Refactoring: break imap_config.dart into smaller files
